requirements.txt
aiohttp==3.10.11
python-dotenv==1.0.1
python-telegram-bot==21.6
pytest==8.3.5
pytest-asyncio==0.24.0
